# Rating average cached in model